Azure Analytic Synapse and Factory Code/ETL Azure Python Code/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-storage-file-datalake
azure-identity
azure-storage-blob
azure-keyvault-secrets
azure-functions
azure-mgmt-storage
pandas
pyarrow
fastparquet
knmy